Moved TokenManagement class to Domain, moved token creation to UserRepository
