+++ AI_NETCORE_API/AI_NETCORE_API/appsettings.Development.json
@@ -4,6 +4,11 @@
       "Default": "Debug",
       "System": "Information",
       "Microsoft": "Information"
+    },
+    "tokenManagement": {
+      "secret": "YourKey-2374-OFFKDI940NG7:56753253-tyuw-5769-0921-kfirox29zoxv",
+      "accessExpiration": 30,
+      "refreshExpiration": 60
     }
   }
 }
