+++ AI_NETCORE_API/Data/Infrastructure/TokenManagement.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace AI_NETCORE_API.Models
+namespace Domain.Infrastructure
 {
     [JsonObject("tokenManagement")]
     public class TokenManagement
