+++ AI_NETCORE_API/Data/Repositories/UserRepo/Concrete/UserRepository.cs
@@ -4,12 +4,20 @@ using Domain.Repositories.UserRepo.Abstract;
 using Data.Models;
 using Domain.DTOToBOConverting;
 using System.Linq;
+using System.IdentityModel.Tokens.Jwt;
+using Microsoft.IdentityModel.Tokens;
+using System.Security.Claims;
+using System;
+using Domain.Infrastructure;
 
 namespace Domain.Repositories.UserRepo.Concrete
 {
-    public class UserRepository: RepositoryBase<User>, IUserRepository
+    public class UserRepository : RepositoryBase<User>, IUserRepository
     {
         private readonly IDTOToBOConverter _converter;
+        private readonly TokenManagement _tokenManagement;
+
+
         public UserRepository(RepositoryContext repositoryContext, IDTOToBOConverter converter)
             : base(repositoryContext)
         {
@@ -33,6 +41,24 @@ namespace Domain.Repositories.UserRepo.Concrete
                 Resources = new List<Resource>()
             });
             RepositoryContext.SaveChanges(true);
+        }
+
+        public string Authenticate(string login, string password)
+        {
+            var tokenHandler = new JwtSecurityTokenHandler();
+            var key = System.Text.Encoding.ASCII.GetBytes(_tokenManagement.Secret);
+            var tokenDescriptor = new SecurityTokenDescriptor
+            {
+                Subject = new ClaimsIdentity(new Claim[]
+                {
+                        new Claim("id", "0"),
+                }),
+                Expires = DateTime.UtcNow.AddDays(7),
+                SigningCredentials = new SigningCredentials(new SymmetricSecurityKey(key), SecurityAlgorithms.HmacSha256Signature)
+            };
+            var token = tokenHandler.CreateToken(tokenDescriptor);
+
+            return tokenHandler.WriteToken(token);
         }
     }
 }
